pixogramFrontEnd/src/app/my-media/my-media.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { RouterTestingModule } from '@angular/router/testing';
import { MyMediaComponent } from './my-media.component';

describe('MyMediaComponent', () => {
  let component: MyMediaComponent;
  let fixture: ComponentFixture<MyMediaComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [
        HttpClientTestingModule,
        FormsModule,
        RouterTestingModule.withRoutes([]),
        ReactiveFormsModule
      ],
      declarations: [ MyMediaComponent ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(MyMediaComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('src is null ',()=>{
    expect(component.src).toBeNull;
  });

  it('images is null ',()=>{
    expect(component.images).toBeNull;
  });
});
